moved enum to constants
