+++ avionics_dash_server/models/assignment_model.py
@@ -1,0 +1,10 @@
+import datetime
+from avionics_dash_server.common.constants import AssignmentType
+
+class Assignment:
+    name: str
+    description: str
+    available: datetime
+    due: datetime
+    points: int
+    assignment_type: AssignmentType
