--- avionics_dash_server/models/enums.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-class AssignmentType(Enum):
-    PERSONAL = 1
-    GROUP = 2
-    DISCUSSION = 3
